requirements.txt
Adafruit-Blinka==8.56.0
adafruit-blinka-displayio==2.1.7
adafruit-circuitpython-bitmap_font==2.3.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-display-text==3.2.4
adafruit-circuitpython-displayio-sh1107==2.0.1
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-sgp30==3.0.11
adafruit-circuitpython-ticks==1.1.2
adafruit-circuitpython-typing==1.11.2
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
flexcache==0.3
flexparser==0.4
obd==0.7.3
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.2
typing_extensions==4.13.2
